requirements/base.txt
Django>=1.10
django-environ>=0.4.0
django-braces>=1.8.1
django-crispy-forms>=1.5.2
django-admin-bootstrapped>=2.5.6
django-authtools>=1.4.0
py-moneyed
django-money
django-phonenumber-field
django-crontab
WhiteNoise
django-dual-authentication